requirements.txt
fastapi
uvicorn
httpx
python-multipart
scikit-learn
pdfminer.six
pytest
